♻️ Refactor: strongly-type EventErrors
